Split helper functions and macros into separate submodules.
